requirements.txt
matplotlib
mne
numpy
torch
torchvision
torch_geometric
pandas
PyYAML
jupyter
scikit-learn
